fix: исправлены баги при переключении листов
